requirements.txt
torch==2.2.1
torchvision==0.17.1
fastapi==0.109.2
python-multipart==0.0.9
pillow==10.2.0
uvicorn==0.27.1
numpy==1.26.4 